hashd: factor out AnonArea into util::anon_area
